+++ croud/cmd.py
@@ -300,6 +300,22 @@ def user_id_or_email_arg(req_args: _ArgumentGroup, opt_args: _ArgumentGroup) -> 
     )
 
 
+def cluster_id_arg(
+    req_args: _ArgumentGroup, opt_args: _ArgumentGroup, required: bool
+) -> None:
+    group = req_args if required else opt_args
+    group.add_argument(
+        "--cluster-id", type=str, help="CrateDB cluster ID", required=required
+    )
+
+
+def product_id_arg(
+    req_args: _ArgumentGroup, opt_args: _ArgumentGroup, required: bool
+) -> None:
+    group = req_args if required else opt_args
+    group.add_argument("--product-id", type=str, help="Filter by product ID.")
+
+
 def product_tier_arg(req_args: _ArgumentGroup, opt_args: _ArgumentGroup) -> None:
     req_args.add_argument("--tier", type=str, help="Product Tier.", required=True)