+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes for croud
 Unreleased
 ==========
 
+- Added ``consumer-sets list`` command that allows users to list their consumer
+  sets and filter them by project ID, cluster ID and product ID.
+
 - Added ``projects users add`` command that allow users with permission
   to add users to projects by specifying a project ID and a user email or
   ID.